Create README.md
# Loopfront AS - Senior React Developer
Loopfront AS offers Norway's largest circular platform for building materials and fixtures. We aim to reduce global material waste and stop the extraction of new natural resources.
As much as 40% of all waste, CO2 emissions, and use of energy and water resources in Norway and Europe comes from the construction industry. Construction waste is one of the biggest pollution sources in Europe, and we help addressing it! The industry itself experiences many challenges associated with the reuse of building materials, including technical, legal, and financial.
Loopfront is an interaction platform that removes these barriers and makes reuse easy. The platform takes care of reuse from A to Z, and helps organizations map, interact, document, and report all available materials for recycling, reuse, and sale. 
If you have the overview, you have control - and this is exactly what we offer our customers! 
As a key member of our Development Team, you will be contributing significantly helping our customers save the environment, as well as saving cost doing so! Would you like to?
Our new colleague should have extensive experience working with the React JavaScript Library (Angular can be an option) and have a desire to contribute to strengthen the circular economy in Norway and Europe.  
| Employment: 100% | Place of work: By agreement | Salary: Fixed | Start: asap |

Main work tasks:
•	Further develop and maintain our Design System (in-house developed)
•	Testing 
•	Perform development tasks as required.
•	Potentially lead our React Team.


Educational requirements:
•	Master (M.Sc.) in Informatics, programming, or related subjects.
•	Alternatively, a Bachelor (B.Sc.) with extensive and relevant experience
Qualifications:
•	Extensive experience working with React – Alternatively Angular
•	Experience with GraphQL
•	Proficient in English Language (Working language is English – Norwegian is considered a plus, but is not a requirement)
It is also beneficial if you know and potentially have experience with any of the following:
•	Figma
•	Tailwind CS
•	Typescript
•	Testing with tools like Cypress and Jest
•	.NET Core / .NET 5
•	Amazon Web Services
•	Docker and container-based architecture
Personal characteristics:
•	You enjoy solving problems and consider them exiting challenges
•	You are innovative, «hands-on» and are driven by progress
•	You enjoy working in cross-functional teams
•	You see potential and opportunities, where others see problems
•	You can communicate fluently in English, both orally and in writing 
Loopfront AS offers:	
•	A unique opportunity to be part of a development-oriented organization with an important social mission 
•	Interesting and varied work tasks in an international environment
•	Good On-Boarding program to get familiar with work tasks and technology
•	Offices in Trondheim, possibly Oslo, and the opportunity to work remotely 
•	Social arrangements (considering the pandemic situation) 
•	Competitive salary and pension scheme

Can you relate to description above? Are you passionate about creating something new, innovative, and sustainable? Please Get in touch or simply apply here
		
Sky rekruttering AS assists Loopfront in this recruitment process, thus please contact our Digital Recruiter Executive - 
Øivind Bergh, per e-mail: oivind@skyrekruttering.no, for any further information, and he will get back to you.